refactor: rename util file, route file; feat: add discoverMedia function in model.ts
